[r8] add r8 and depot_tools submodules.
depot_tools is required to build r8. On Windows it is however not used but
a binary executable zip will be downloaded and used instead.
(see https://commondatastorage.googleapis.com/chrome-infra-docs/flat/depot_tools/docs/html/depot_tools_tutorial.html#_setting_up for details)

r8 build is not integrated in the primary Makefile and .sln yet.

(r8 is not a "build tool" but we build msbuild tasks dependencies there too.)