Requested changes from review + build refactoring
~~ xa-prep-tasks ~~

The `<DownloadUri/>` MSBuild task now has an optional `HashHeader`
property:
- If set, a http HEAD request is made to the URL, and the destination
  file path gets a suffix added for the value.
- To make this work properly, `DestinationFiles` is also an `[Output]`
  property, since these paths could change as a result.

~~ android-toolchain ~~

Chromium's `depot_tools` are downloaded the same as all of our other
dependencies. I also updated all `<DownloadUri/>` calls to specify
`HashHeader` and `DestinationFiles` as an `<Output/>`

While implementing this, I became quite frustrated waiting on all of
the other zip files to get deleted and recreated... So I went ahead
and split up these operations into four targets:

- `_UnzipAndroidSdkItems` with `@(_AndroidSdkItemDownloads)` as inputs
- `_UnzipAndroidNdkItems` with `@(_AndroidNdkItemDownloads)` as inputs
- `_UnzipAntItems` with `@(_AntItemDownloads)` as inputs
- `_UnzipChromeItems` with `@(_ChromeItemDownloads)` as inputs

In the future, if these downloads change, *all* of the directories
won't need to be nuked and extracted again. I did this in a way so
that the existing stamp files will be preserved, so after this change
is merged, locally the build on your machine will only need to download
`depot_tools` and extract it if you have a reasonably recent
`~/android-toolchain` directory.

Other changes:

- Fixed formatting
- Removed usage of the obsolete `<CreateItem/>` MSBuild task
- `_UnzipChromeItems` also "bootstraps" `depot_tools`, and has to
  set/unset `$PATH` appropriately. I did this in a similar manner as
  the license acceptance for the Android SDK.

~~ r8.targets ~~

- Removed all of the `depot_tools` downloading code, and setup
  `android-toolchain` as a `@(ProjectReference)`
- All targets that set `$PATH`, unset it when finished
- Use `$(_GradleArgs)` to supply `--no-daemon`
- `_CleanR8` should delete `*.jar` files in our build tree

~~ Xamarin.Android.Build.Tasks ~~

- Fixed any formatting that we noticed
- Renamed `$(AndroidD8JarPath)` and `$(AndroidR8JarPath)` to be
  prefixed with `Android`
- Refactored `$(IntermediateOutputPath)_dex_stamp` stamp file in
  `_CompileToDalvikWithDx` and `_CompileToDalvikWithD8` to match the
  new convention of `$(_AndroidStampDirectory)_CompileToDalvik.stamp`
- `*.dex` files weren't in `FileWrites`???
- `CompileToDalvik` had a `DexOutputs` output property that was
  completely unused, so I removed it. Also removed extra log messages.