[Xamarin.Android.Build.Tasks] more stamp files + refactoring
~~ _CreateIntermediateAssembliesDir ~~

This target has `Inputs` and `Outputs`, but all it does it create a
directory? I moved the `<MakeDir/>` call into the
`_CreatePackageWorkspace` target, which was in the same dependency
chain.

I also added `!Exists` check to all the `<MakeDir/>` calls in
`_CreatePackageWorkspace`, as I found there is a minor performance
benefit from this in the past.

~~ _CopyConfigFiles ~~

This target uses a list of `*.config` files as an output, but the
`<CopyConfigFiles/>` has logic similar to `<CopyIfChanged/>`. I
removed the `<CopyConfigFiles/>` MSBuild task, as it seemed
duplicative of `<CopyIfChanged/>`. I also switched to using a "stamp"
file following our convention.

TODO MDB/PDB changes