[Xamarin.Android.Build.Tasks] use SRM ResolveLibraryProjectImports
Context: https://github.com/dotnet/corefx/tree/master/src/System.Reflection.Metadata/src/System/Reflection/Metadata
Context: https://github.com/jonathanpeppers/Benchmarks

There is a new System.Reflection.Metadata library from corefx for
reading .NET assemblies. It is a bit more performant than Mono.Cecil
because it is a different library with different opinions:

- SRM has a forward "reader" style API
- SRM uses lots of structs, and you have to do an additional call to
  lookup strings generally.
- SRM, as far as I have seen, doesn't have APIs to modify and write
  out new assemblies.

From a simple benchmark, it seems that SRM is 10x faster on
Windows/.NET framework and 5x faster on macOS/Mono.

So it makes sense for use to use SRM when reading assemblies, and
continue with Mono.Cecil for the linker and other things that modify
assemblies.

There are a few places we can take advantage of SRM, but the simplest
to start with was `ResolveLibraryProjectImports`: