Issue with readonly obj\Debug\android\bin\class\**\*.class files?
Context: http://feedback.devdiv.io/443423

This was the idea of what I thought was wrong. But it is working for
me! So readonly files aren't the problem...

The error reported is:

    Unable to remove directory "obj\Debug\81\android\bin\classes".
    Could not find a part of the path 'AccessibilityManagerCompat_TouchExplorationStateChangeListenerImplementor.class'.

`Could not find a part of the path XYZ` is the same error I get in
powershell if I try to make a directory over the `MAX_PATH` limit:

    > $env:HOME > AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA.txt
    out-file : Could not find a part of the path 'C:\full\path\to\AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA.txt'.
    At line:1 char:1
    + $env:HOME > AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA ...
    + ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
        + CategoryInfo          : OpenError: (:) [Out-File], DirectoryNotFoundException
        + FullyQualifiedErrorId : FileOpenFailure,Microsoft.PowerShell.Commands.OutFileCommand

This makes me think we have a file in
`obj\Debug\81\android\bin\classes` over the `MAX_PATH` limit.

Unsure how to fix yet?