[Xamarin.Android.Build.Tasks] prototype of a Java "daemon"
This creates a new mode for `JavaToolTask`, where if `UseDaemon=True`:

* Instead of just running a new `java.exe`, a new `JavaDaemonClient`
  class will start a java process that runs `java-daemon.jar`.
* The daemon reads stdin to run a jar file's `Main` method, and
  replies to stdout with the result of running that jar file.
* If `JavaDaemonClient` is disposed (and when the MSBuild node shuts
  down), the java process will exit.

~~ Results ~~

I built the Xamarin.Forms project in this repo, using
`%MSBUILDNOINPROCNODE%` = 1.

The command was:

    git clean -dxf tests
    xabuild tests\Xamarin.Forms-Performance-Integration\Droid\Xamarin.Forms.Performance.Integration.Droid.csproj /r /t:SignAndroidPackage /bl /clp:performancesummary /p:AndroidUseJavaDaemon=True

Results:

    Before:
    3532 ms  AndroidApkSigner                           1 calls
    5104 ms  D8                                         1 calls
    Time Elapsed 00:00:27.20

    After (daemon running already):
    2651 ms  AndroidApkSigner                           1 calls
    3211 ms  D8                                         1 calls
    Time Elapsed 00:00:20.56

This seems like it saves ~7 seconds for this app. In both cases an
MSBuild node was already running.

Originally `java.exe` was using ~531MB of RAM on my PC that has a
total of 64GB of RAM. For comparison `MSBuild.exe` is using ~62MB.

I added a call to `System.gc()` and used the `-XX:+UseG1GC` switch to
get `java.exe` idling at ~90MB.

https://www.javacodegeeks.com/2017/11/minimize-java-memory-usage-right-garbage-collector.html