[Xamarin.Android.build.Tasks] WIP AOT can now be used with Fast Dev
Context: https://github.com/xamarin/xamarin-android/issues/3828

Moved the `<Aot/>` MSBuild task call from `_BuildApkEmbed` to `_PrepareBuildApk`. This allows it to work for Fast Deployment where `_BuildApkFastDev` is used.

NOTE: this isn't complete, as AOT will need to be reworked to be incremental. This change runs AOT on every build, but was enough to test.