[WIP] <ResolveAssemblies/> remove NuGet & recursive lookup
One of the remaining pain points in the Xamarin.Android build is the
`<ResolveAssemblies/>` MSBuild task.

1. It isn't particularly fast and runs on *every* build. During the
   devloop it runs twice for `Build` and `Install`. It runs for DTBs,
   designer builds, too.

2. It breaks for certain NuGet packages as mentioned on a few Github
   issues:

https://github.com/xamarin/xamarin-android/issues/3898
https://github.com/xamarin/xamarin-android/issues/4074

After some research and looking at build logs, we may be able to
rework the behavior:

1. Remove NuGet-related code and recursive assembly searching.

2. Use `@(_ReferencePath)` and `@(_ReferenceDependencyPaths)`.

3. Include assemblies even if they have `%(ResolvedFrom)` ==
   `ImplicitlyExpandDesignTimeFacades`.

With these changes I am still able to build SmartHotel360, and the
`<ResolveAssemblies/>` is an order of magnitude quicker:

    Before:
    310 ms  ResolveAssemblies                          1 calls
    After:
     44 ms  ResolveAssemblies                          1 calls

So this would save ~532ms (x2 620ms -> 88ms) from the dev-loop.

This is a big WIP, as I need to figure out what types of projects this
might break. Let's see what happens on CI!