[Xamarin.Android.Build.Tasks] introduce DirectoryAssemblyResolverTask
This introduces a new `DirectoryAssemblyResolverTask` that enables a
`DirectoryAssemblyResolver` to be shared throughout a build using
`RegisterTaskObject`. It can be a common base class for
`<GenerateJavaStubs/>` and `<GenerateTypeMaps/>`.

* `GetResolver` will retrieve a cached resolver, creating a new one if
  needed.

* `GetJavaTypes` will retrieve a cached `List<TypeDefinition>` via
  `JavaTypeScanner`.

* `Assemblies` will contain a `List<string>` of all loaded assemblies.

Results of an incremental build of the Xamarin.Forms integration
project after changing it to use `ProduceReferenceAssembly=False`:

    master:
    GenerateJavaStubs = 392 ms

    this branch before:
    GenerateTypeMaps = 570 ms
    GenerateJavaStubs = 284 ms

    this branch after:
    GenerateTypeMaps = 242 ms
    GenerateJavaStubs = 162 ms

One note is that `<GenerateJavaStubs/>` runs last, so it will
currently make the call to `UnregisterAll`. If we for some reason
*forgot* to call this, the `DirectoryAssemblyResolver` would get
disposed by the MSBuild infrastructure. `RegisterTaskObject` handles
`IDisposable`.