[Xamarin.Android.Build.Tasks] clean $(IntermediateOutputPath) if build.props changes
I found a crash at runtime, if you:

1. Build / Run a project with 16.4

2. Build / Run the same project with 16.6

The stack trace is:

    java.lang.UnsatisfiedLinkError: dlopen failed: cannot locate symbol "jm_typemap" referenced by "/data/app/com.companyname.app59-bgIEcmaPpoMxpR948Wo8ZA==/lib/x86/libmonodroid.so"...
        at java.lang.Runtime.loadLibrary0(Runtime.java:1071)
        at java.lang.Runtime.loadLibrary0(Runtime.java:1007)
        at java.lang.System.loadLibrary(System.java:1667)
        at mono.MonoPackageManager.LoadApplication(MonoPackageManager.java:80)
        at mono.MonoRuntimeProvider.attachInfo(MonoRuntimeProvider.java:48)
        at android.app.ActivityThread.installProvider(ActivityThread.java:6983)
        at android.app.ActivityThread.installContentProviders(ActivityThread.java:6528)
        at android.app.ActivityThread.handleBindApplication(ActivityThread.java:6445)
        at android.app.ActivityThread.access$1300(ActivityThread.java:219)
        at android.app.ActivityThread$H.handleMessage(ActivityThread.java:1859)
        at android.os.Handler.dispatchMessage(Handler.java:107)
        at android.os.Looper.loop(Looper.java:214)
        at android.app.ActivityThread.main(ActivityThread.java:7356)
        at java.lang.reflect.Method.invoke(Native Method)
        at com.android.internal.os.RuntimeInit$MethodAndArgsCaller.run(RuntimeInit.java:492)
        at com.android.internal.os.ZygoteInit.main(ZygoteInit.java:930)

What happened is in ce2bc689, the `jm_typemap` and `mj_typemap`
symbols were removed. But we don't actually have anything that cleans
the generated assembly files when `$(XamarinAndroidVersion)` changes.

We have a `DeleteBinObjTest` test, why didn't that catch this?

Apparently the test was triggering:

    Building target "_CleanIntermediateIfNuGetsChange" completely.
    Input file "HelloForms.Android\obj\project.assets.json" is newer than output file "obj\Debug\90\stamp\_CleanIntermediateIfNuGetsChange.stamp".

And so the test wasn't actually *testing* much of anything!

As part of that test, we have to run a NuGet restore. The zip file
contains a project that was restored on another machine. We *have* to
run `/t:Restore` for things to build. The changes in this PR remove
`_CleanIntermediateIfNuGetsChange`, so we should be OK going forward.

I also updated the test to include projects built with 16.4 and verify
the app launches.

~~ Actual Changes ~~

To fix the real problem here, I think we need to fundamentally change
what happens when values in `build.props` /
`$(_AndroidBuildPropertiesCache)` changes. Currently we are using it
to trigger several MSBuild targets, but I believe we should also use
it to trigger the `_CleanMonoAndroidIntermediateDir` target if the
file changes.

1. Repurpose `_CleanIntermediateIfNuGetsChange` to instead be a
   `_CleanIntermediateIfNeeded` target that does not run on
   design-time builds.

2. Include the `GetLastWriteTime` value of `$(_NuGetAssetsFile)` in
   `build.props`.

3. Remove `DesignTimeBuild=$(DesignTimeBuild)` from `build.props`. We
   use a different path for DTBs at `obj\Debug\designtime\build.props`
   anyway.

4. `_CleanMonoAndroidIntermediateDir` can't delete `build.props`
   anymore. I moved this so it will still happen during a full
   `Clean`.

5. Remove the `_CleanIntermediateIfPackageNamingPolicyChanges` and
   `_CleanupOldStaticResources` MSBuild targets. They are no longer
   needed at all.

6. Remove `Xamarin.Android.DefaultOutputPaths.targets`.

This also partially reverts 13d216fd, as we don't need a
`obj\Debug\100` directory anymore. When `$(TargetFrameworkVersion)`
changes, a partial clean will occur. This is nice, since it will buy
us a few more characters for `%MAX_PATH%`. There was also some logic
that prevented the extra directory on macOS. Now Windows and macOS
will have the same `$(IntermediateOutputPath)`.