[xabt] Compute `$DOTNET_DiagnosticPorts` using MSBuild properties (#10351)
If you run `dotnet-trace collect --dsrouter android`, it says:

    Start an application on android device with ONE of the following environment variables set:
    [Default Tracing]
    DOTNET_DiagnosticPorts=127.0.0.1:9000,nosuspend,connect
    [Startup Tracing]
    DOTNET_DiagnosticPorts=127.0.0.1:9000,suspend,connect

Setting `$DOTNET_DiagnosticPorts` is non-trivial, so as a step to
simplify this, we are adding multiple MSBuild properties to configure
this value.

This would allow the log message to say:

    Build and run an Android application such as:
    [Default Tracing]
    dotnet build -t:Run -c Release -p:DiagnosticAddress=127.0.0.1 -p:DiagnosticPort=9000 -p:DiagnosticSuspend=false -p:DiagnosticListenMode=connect
    [Startup Tracing]
    dotnet build -t:Run -c Release -p:DiagnosticAddress=127.0.0.1 -p:DiagnosticPort=9000 -p:DiagnosticSuspend=true -p:DiagnosticListenMode=connect

You could also set `$(DiagnosticConfiguration)`, but you would need to
escape the `,` character with `%2c`.

Setting any of the new properties also implicitly means that
`$(AndroidEnableProfiler)` is set to `true`.